// web/lib/api/generated/model/topTenant.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export interface TopTenant {
  tenant_id: string;
  name: string;
  orders: number;
  /** @pattern ^-?\d{0,14}(?:\.\d{0,2})?$ */
  value: string;
}